// src/shared/api/soil-lab/model/testResultCreate.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * soil_laboratory
 * OpenAPI spec version: 0.1.0
 */
import type { TestResultCreateMeasurements } from './testResultCreateMeasurements'

export interface TestResultCreate {
  sampleId: string
  parameterId: string
  measurements?: TestResultCreateMeasurements
  context: string
}