// src/shared/api/soil-lab/model/testResultDetailResponse.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * soil_laboratory
 * OpenAPI spec version: 0.1.0
 */
import type { MeasurementShortResponse } from './measurementShortResponse'
import type { ParameterShortResponse } from './parameterShortResponse'
import type { SampleShortResponse } from './sampleShortResponse'
import type { TestResultDetailResponseCreatedById } from './testResultDetailResponseCreatedById'
import type { TestResultDetailResponseDeletedAt } from './testResultDetailResponseDeletedAt'
import type { TestResultDetailResponseLowerLimit } from './testResultDetailResponseLowerLimit'
import type { TestResultDetailResponseMeanValue } from './testResultDetailResponseMeanValue'
import type { TestResultDetailResponseUpdatedById } from './testResultDetailResponseUpdatedById'
import type { TestResultDetailResponseUpperLimit } from './testResultDetailResponseUpperLimit'
import type { TestResultDetailResponseVariationPercentage } from './testResultDetailResponseVariationPercentage'

export interface TestResultDetailResponse {
  deletedAt?: TestResultDetailResponseDeletedAt
  createdAt: string
  createdById?: TestResultDetailResponseCreatedById
  updatedAt: string
  updatedById?: TestResultDetailResponseUpdatedById
  id: string
  sample: SampleShortResponse
  parameter: ParameterShortResponse
  measurements: MeasurementShortResponse[]
  meanValue: TestResultDetailResponseMeanValue
  variationPercentage: TestResultDetailResponseVariationPercentage
  lowerLimit: TestResultDetailResponseLowerLimit
  upperLimit: TestResultDetailResponseUpperLimit
  isCompliant: boolean
  readonly isDeleted: boolean
}